Some refactors Bench: 6189429
